pyproject.toml
[project]
name = "okulib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.121.0",
    "google-genai>=1.49.0",
    "langchain>=1.0.4",
    "langchain-text-splitters>=1.0.0",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]
